requirement.txt
fastapi
uvicorn
pydantic
langchain
sentence-transformers
faiss-cpu
pandas